3916 Moving FormState to formStateTypes.ts
