+++ src/components/documentBuilder/hooks/useDeleteElement.ts
@@ -16,7 +16,7 @@
  */
 
 import { getIn, useFormikContext } from "formik";
-import { FormState } from "@/pageEditor/pageEditorTypes";
+import { FormState } from "@/pageEditor/extensionPoints/formStateTypes";
 import getElementCollectionName from "@/components/documentBuilder/edit/getElementCollectionName";
 import { produce } from "immer";
 import { produceExcludeUnusedDependencies } from "@/components/fields/schemaFields/serviceFieldUtils";
