+++ src/pageEditor/components/UrlPatternWidget.tsx
@@ -20,7 +20,7 @@ import { useFormikContext } from "formik";
 import ArrayWidget from "@/components/fields/schemaFields/widgets/ArrayWidget";
 import FieldRuntimeContext from "@/components/fields/schemaFields/FieldRuntimeContext";
 import { PAGE_EDITOR_DEFAULT_BRICK_API_VERSION } from "@/pageEditor/extensionPoints/base";
-import { FormState } from "@/pageEditor/pageEditorTypes";
+import { FormState } from "@/pageEditor/extensionPoints/formStateTypes";
 import { SchemaFieldProps } from "@/components/fields/schemaFields/propTypes";
 import { Schema } from "@/core";
