+++ src/pageEditor/tabs/editTab/editHelpers.ts
@@ -20,7 +20,7 @@ import { freshIdentifier } from "@/utils";
 import { selectReaderIds } from "@/blocks/readers/readerUtils";
 import getType from "@/runtime/getType";
 import { BlockType } from "@/runtime/runtimeTypes";
-import { FormState } from "@/pageEditor/pageEditorTypes";
+import { FormState } from "@/pageEditor/extensionPoints/formStateTypes";
 import { traversePipeline } from "@/pageEditor/utils";
 import { BlockPipeline } from "@/blocks/types";
 import { PipelineMap } from "@/pageEditor/uiState/uiStateTypes";
